+++ src/alfanous/Support/PyArabic/__init__.py
