+++ src/alfanous/Support/PyArabic/ar_ctype.py
