+++ src/alfanous/Support/PyArabic/arabic_const.py
