+++ src/alfanous/Support/PyArabic/araby.py
