+++ src/alfanous/Support/PyArabic/araby_constants.py
