+++ src/alfanous/Support/PyArabic/araby_predicates.py
